aanpassingen backend + testen + angularjs frontend, deelname toegevoegd + controle op deelname, lijst cirkelsessies aangepast

+++ src/main/java/be/kdg/teamh/services/DeelnameServiceImpl.java
@@ -1,8 +1,15 @@
 package be.kdg.teamh.services;
 
+import be.kdg.teamh.entities.Cirkelsessie;
 import be.kdg.teamh.entities.Deelname;
+import be.kdg.teamh.entities.Gebruiker;
+import be.kdg.teamh.exceptions.AlreadyJoinedCirkelsessie;
+import be.kdg.teamh.exceptions.CirkelsessieNotFound;
 import be.kdg.teamh.exceptions.DeelnameNotFound;
+import be.kdg.teamh.exceptions.GebruikerNotFound;
+import be.kdg.teamh.repositories.CirkelsessieRepository;
 import be.kdg.teamh.repositories.DeelnameRepository;
+import be.kdg.teamh.repositories.GebruikerRepository;
 import be.kdg.teamh.services.contracts.DeelnameService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,30 +19,66 @@ import java.util.List;
 
 @Service
 @Transactional
-public class DeelnameServiceImpl implements DeelnameService
-{
+public class DeelnameServiceImpl implements DeelnameService {
     @Autowired
     private DeelnameRepository repository;
 
+    @Autowired
+    private GebruikerRepository gebruikerRepository;
+
+    @Autowired
+    private CirkelsessieRepository cirkelsessieRepository;
+
     @Override
-    public List<Deelname> all()
-    {
+    public List<Deelname> all() {
         return repository.findAll();
     }
 
     @Override
-    public void create(Deelname deelname)
-    {
-        repository.save(deelname);
+    public void create(int id, int userId) throws DeelnameNotFound, GebruikerNotFound, CirkelsessieNotFound, AlreadyJoinedCirkelsessie {
+
+        Cirkelsessie cirkelsessie = cirkelsessieRepository.findOne(id);
+        Gebruiker gebruiker = gebruikerRepository.findOne(userId);
+
+
+        for (Deelname deelname : repository.findAll()) {
+            if (deelname.getGebruiker().getId() == gebruiker.getId() && deelname.getCirkelsessie().getId() == cirkelsessie.getId()) {
+                throw new AlreadyJoinedCirkelsessie();
+
+            }
+        }
+
+        if (cirkelsessie == null) {
+            throw new CirkelsessieNotFound();
+        }
+
+        if (gebruiker == null) {
+            throw new GebruikerNotFound();
+        }
+
+        //deelname
+        Deelname deelname = new Deelname(0, false);
+        deelname.setCirkelsessie(cirkelsessie);
+        deelname.setGebruiker(gebruiker);
+        Deelname savedDeelname = repository.save(deelname);
+
+        gebruiker.addDeelname(savedDeelname);
+        cirkelsessie.addDeelname(savedDeelname);
+
+        //gebruiker
+        gebruikerRepository.saveAndFlush(gebruiker);
+
+        //cirkelsessie
+        cirkelsessieRepository.saveAndFlush(cirkelsessie);
+
+
     }
 
     @Override
-    public Deelname find(int id) throws DeelnameNotFound
-    {
+    public Deelname find(int id) throws DeelnameNotFound {
         Deelname deelname = repository.findOne(id);
 
-        if (deelname == null)
-        {
+        if (deelname == null) {
             throw new DeelnameNotFound();
         }
 
@@ -43,8 +86,7 @@ public class DeelnameServiceImpl implements DeelnameService
     }
 
     @Override
-    public void update(int id, Deelname deelname) throws DeelnameNotFound
-    {
+    public void update(int id, Deelname deelname) throws DeelnameNotFound {
         Deelname old = find(id);
 
         old.setCirkelsessie(deelname.getCirkelsessie());
@@ -56,10 +98,12 @@ public class DeelnameServiceImpl implements DeelnameService
     }
 
     @Override
-    public void delete(int id) throws DeelnameNotFound
-    {
+    public void delete(int id) throws DeelnameNotFound {
         Deelname deelname = find(id);
 
+        if (deelname == null) {
+            throw new DeelnameNotFound();
+        }
         repository.delete(id);
     }
 }
